Fix screenshare feature (#51)
This wasn't working, might have broken within the last version or two of
xcode. I saw that the extension didn't appear in the list, and once I
fixed that it didn't work due the main app not having access to the app
group container to create the socket file. Fixed by:

- added the app group to ios entitlements via signing & capabilities
page
- embed broadcast extension upon install
- add bundle id and app group keys to plist entry in project file